split mock_test.go from data_node_test.go
